refactor(GN-73): 타입 분리
