+++ eslint.config.mjs
@@ -57,6 +57,10 @@ export default defineConfig([
           index: true,
           rules: [
             {
+              case: 'camel',
+              target: '**/components/**/*.ts',
+            },
+            {
               case: 'pascal',
               target: '**/components/**',
             },
