+++ src/constants/modalConstants.ts
@@ -1,13 +1,11 @@
-// 기본 메시지
 export const DEFAULT_MESSAGES = {
-  ALERT: '메시지가 없습니다.',
+  ALERT: '알림이 있습니다.',
   CONFIRM: '예약을 취소하시겠습니까?',
   ERROR: '오류가 발생했습니다.',
-} as const;
+};
 
-// 버튼 텍스트
 export const BUTTON_TEXTS = {
   CONFIRM: '확인',
   YES: '취소하기',
   NO: '아니오',
-} as const;
+};
